+++ docs/reference/types/7-file-types.md
@@ -20,3 +20,4 @@
     options:
         members:
             - info
+            - open
